--- src/main/java/com/example/mate/domain/goods/repository/GoodsReviewRepository.java
@@ -1,11 +1,0 @@
-package com.example.mate.domain.goods.repository;
-
-import com.example.mate.domain.goods.entity.GoodsReview;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface GoodsReviewRepository extends JpaRepository<GoodsReview, Long> {
-
-    int countByRevieweeId(Long revieweeId);
-
-    boolean existsByGoodsPostIdAndReviewerId(Long goodsPostId, Long reviewerId);
-}
