+++ src/main/java/com/example/mate/domain/goodsChat/dto/response/GoodsChatRoomResponse.java
@@ -2,7 +2,7 @@ package com.example.mate.domain.goodsChat.dto.response;
 
 import com.example.mate.common.response.PageResponse;
 import com.example.mate.domain.constant.TeamInfo;
-import com.example.mate.domain.goods.entity.GoodsPost;
+import com.example.mate.domain.goodsPost.entity.GoodsPost;
 import com.example.mate.domain.goodsChat.entity.GoodsChatRoom;
 import lombok.Builder;
 import lombok.Getter;
