+++ src/main/java/com/example/mate/domain/goodsChat/entity/GoodsChatPart.java
@@ -1,6 +1,6 @@
 package com.example.mate.domain.goodsChat.entity;
 
-import com.example.mate.domain.goods.entity.Role;
+import com.example.mate.domain.goodsPost.entity.Role;
 import com.example.mate.domain.member.entity.Member;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
