+++ src/main/java/com/example/mate/domain/goodsPost/entity/Role.java
@@ -1,0 +1,14 @@
+package com.example.mate.domain.goodsPost.entity;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum Role {
+
+    SELLER("판매자"),
+    BUYER("구매자");
+
+    private final String value;
+}
