+++ src/main/java/com/example/mate/domain/goodsPost/entity/Status.java
@@ -1,0 +1,14 @@
+package com.example.mate.domain.goodsPost.entity;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum Status {
+
+    OPEN("거래중"),
+    CLOSED("거래완료");
+
+    private final String value;
+}
