+++ src/main/java/com/example/mate/domain/goodsPost/repository/GoodsPostRepositoryCustom.java
@@ -1,0 +1,11 @@
+package com.example.mate.domain.goodsPost.repository;
+
+import com.example.mate.domain.goodsPost.entity.Category;
+import com.example.mate.domain.goodsPost.entity.GoodsPost;
+import com.example.mate.domain.goodsPost.entity.Status;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+
+public interface GoodsPostRepositoryCustom {
+    Page<GoodsPost> findPageGoodsPosts(Long teamId, Status status, Category category, Pageable pageable);
+}
