+++ src/main/java/com/example/mate/domain/goodsReview/entity/GoodsReview.java
@@ -1,0 +1,55 @@
+package com.example.mate.domain.goodsReview.entity;
+
+import com.example.mate.common.BaseTimeEntity;
+import com.example.mate.domain.constant.Rating;
+import com.example.mate.domain.goodsPost.entity.GoodsPost;
+import com.example.mate.domain.member.entity.Member;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "goods_review")
+@Getter
+@Builder
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class GoodsReview extends BaseTimeEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "post_id", nullable = false)
+    private GoodsPost goodsPost;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "reviewer_id", nullable = false)
+    private Member reviewer;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "reviewee_id", nullable = false)
+    private Member reviewee;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private Rating rating;
+
+    @Column(name = "review_content", columnDefinition = "TEXT")
+    private String reviewContent;
+}
