+++ src/main/java/com/example/mate/domain/goodsReview/repository/GoodsReviewRepositoryCustom.java
@@ -1,0 +1,10 @@
+package com.example.mate.domain.goodsReview.repository;
+
+import com.example.mate.domain.member.dto.response.MyReviewResponse;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+
+public interface GoodsReviewRepositoryCustom {
+
+    Page<MyReviewResponse> findGoodsReviewsByRevieweeId(Long memberId, Pageable pageable);
+}
