+++ src/main/java/com/example/mate/domain/member/service/MemberService.java
@@ -7,9 +7,9 @@ import com.example.mate.common.security.util.JwtUtil;
 import com.example.mate.domain.constant.TeamInfo;
 import com.example.mate.domain.file.FileService;
 import com.example.mate.domain.file.FileValidator;
-import com.example.mate.domain.goods.entity.Status;
-import com.example.mate.domain.goods.repository.GoodsPostRepository;
-import com.example.mate.domain.goods.repository.GoodsReviewRepository;
+import com.example.mate.domain.goodsPost.entity.Status;
+import com.example.mate.domain.goodsPost.repository.GoodsPostRepository;
+import com.example.mate.domain.goodsReview.repository.GoodsReviewRepository;
 import com.example.mate.domain.mate.repository.MateReviewRepository;
 import com.example.mate.domain.mate.repository.VisitPartRepository;
 import com.example.mate.domain.member.dto.request.JoinRequest;
