+++ src/main/java/com/example/mate/domain/member/service/ProfileService.java
@@ -3,10 +3,10 @@ package com.example.mate.domain.member.service;
 import com.example.mate.common.error.CustomException;
 import com.example.mate.common.error.ErrorCode;
 import com.example.mate.common.response.PageResponse;
-import com.example.mate.domain.goods.entity.GoodsPost;
-import com.example.mate.domain.goods.entity.Status;
-import com.example.mate.domain.goods.repository.GoodsPostRepository;
-import com.example.mate.domain.goods.repository.GoodsReviewRepositoryCustom;
+import com.example.mate.domain.goodsPost.entity.GoodsPost;
+import com.example.mate.domain.goodsPost.entity.Status;
+import com.example.mate.domain.goodsPost.repository.GoodsPostRepository;
+import com.example.mate.domain.goodsReview.repository.GoodsReviewRepositoryCustom;
 import com.example.mate.domain.match.entity.Match;
 import com.example.mate.domain.mate.entity.MateReview;
 import com.example.mate.domain.mate.repository.MateReviewRepository;
